mcp/damodaran_valuation/pyproject.toml
[project]
name = "damodaran-valuation"
version = "0.1.0"
description = "Damodaran MCP Valuation Server"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "pandas>=2.2.0",
    "psycopg[binary]>=3.2.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.0",
    "xlrd>=2.0.1",
]

[tool.uv]
package = true

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

